pyproject.toml
[project]
name = "ppe-detection-system"
version = "0.1.0"
description = "Construction Site PPE Detection System with FastAPI backend and Gradio frontend"
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
    # Deep Learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.3.0",

    # Web API & Interface
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "gradio>=4.0.0",

    # Image Processing
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",

    # Utilities
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
ppe-api = "src.main:main"
ppe-gradio = "src.app_gradio:main"

[tool.hatch.build.targets.wheel]
include = ["src/*.py"]
